add layer param bases and fix operations between different bases
