+++ src/xlnstorch/operators/__init__.py
@@ -87,9 +87,6 @@ from .internal_lns_ops import (
     lns_conv2d,
     lns_conv3d,
 )
-from ._function import (
-    LNSFunction,
-)
 from .arithmetic_ops import (
     implement_sbdb,
     sbdb,
@@ -122,8 +119,6 @@ __all__ = [
     "lns_prod",
     "lns_matmul",
     "lns_transpose",
-
-    "LNSFunction",
 
     "lns_equal",
     "lns_eq",
