+++ src/xlnstorch/operators/loss_ops.py
@@ -1,9 +1,8 @@
 import math
 
 import torch
-from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, zeros_like
+from .. import LNS_ZERO, LNSTensor, lnstensor, format_lnstensor_operands, implements, zeros_like, LNSFunction
 from . import(
-    LNSFunction,
     lns_sub,
     lns_mul,
     lns_div,
