+++ src/xlnstorch/tensor.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import Any, Union, List
+from typing import Any, Union, List, Tuple
 
 import math
 import numpy as np
@@ -711,7 +711,90 @@ class LNSTensor:
     def sigmoid(self):
         return torch.sigmoid(self)
 
-class FloatToLNS(torch.autograd.Function):
+class LNSFunction(torch.autograd.Function):
+    """
+    Base class for LNS operations that require custom forward and backward methods.
+    This class should be subclassed for specific LNS operations.
+    """
+
+    @staticmethod
+    def forward(ctx, *args, **kwargs):
+        """
+        Forward pass for the LNS operation.
+        Should be implemented in subclasses.
+        """
+        raise NotImplementedError("Forward method must be implemented in subclasses.")
+
+    @staticmethod
+    def backward(ctx, *grad_outputs):
+        """
+        Backward pass for the LNS operation.
+        Should be implemented in subclasses.
+        """
+        raise NotImplementedError("Backward method must be implemented in subclasses.")
+
+    @classmethod
+    def apply(cls, *args, **kwargs):
+        """
+        Applies the LNS operation defined by this class.
+        This method is used to call the forward and backward methods.
+
+        Note that any keyword arguments passed to this method will raise an error,
+        as `torch.autograd.Function` does not support keyword arguments. Instead,
+        use positional arguments only.
+
+        In addition, this method converts any `LNSTensor` arguments to their
+        internal representation (i.e., the underlying tensor) before calling the
+        forward method. This is necessary for LNSTensor internal behavior.
+        """
+        # This check is also performed in the base class, but we do it here too
+        # in case PyTorch decides to change the behavior of the apply method.
+        if kwargs:
+            raise ValueError("torch.autograd.Function does not support keyword arguments. Please use positional arguments only.")
+
+        # Convert LNSTensor arguments to internal representation. This is necessary
+        # because the autograd.Function expects tensors, not LNSTensor objects.
+        internal_args = []
+        for arg in args:
+            if isinstance(arg, LNSTensor):
+                internal_args.append(arg._lns)
+            else:
+                internal_args.append(arg)
+
+        # call the forward method of the class with the internal arguments
+        result = super().apply(*internal_args)
+
+        # get all output tensors and store them in a tuple
+        if isinstance(result, torch.Tensor):
+            result_tuple = (result,)
+        elif isinstance(result, (tuple, list)):
+            result_tuple = tuple(result)
+        else:
+            result_tuple = tuple()
+
+        # we register hooks on each input to each output for gradient accumulation
+        for output in result_tuple:
+
+            # only register hooks for tensors outputs that require gradients
+            if not (isinstance(output, torch.Tensor) and output.requires_grad):
+                continue
+
+            # get the gradient edge for the output tensor.
+            edge = torch.autograd.graph.get_gradient_edge(output)
+
+            for i in range(len(args)):
+
+                # only register hooks for LNSTensor inputs with gradients
+                if not (isinstance(args[i], LNSTensor) and args[i].requires_grad):
+                    continue
+
+                # track operation by registering a hook on the input tensor
+                # to use custom addition logic each time it receives a gradient
+                args[i]._track_operation(edge, i)
+
+        return result
+
+class FloatToLNS(LNSFunction):
 
     @staticmethod
     def forward(x, base):
@@ -740,6 +823,120 @@ class FloatToLNS(torch.autograd.Function):
         sign = torch.where((packed_grad_output & 1).bool(), -1.0, 1.0)
 
         return torch.where(torch.eq(packed_grad_output | 1, LNS_ZERO), 0.0, sign * torch.pow(base, exponent)), None
+
+class LNSChangeBaseFunction(LNSFunction):
+
+    @staticmethod
+    def forward(tensor, old_base, new_base):
+        packed_int = tensor.to(torch.int64)
+        sign_bit = packed_int & 1
+        exponent = (packed_int >> 1).to(torch.float64)
+
+        exponent_new = exponent * torch.log(old_base) / torch.log(new_base)
+        new_packed_int = (exponent_new.round().to(torch.int64) << 1) | sign_bit
+        new_tensor = new_packed_int.to(torch.float64)
+
+        return new_tensor
+
+    @staticmethod
+    def setup_context(ctx, inputs, outputs):
+        _, old_base, new_base = inputs
+        ctx.save_for_backward(old_base, new_base)
+
+    @staticmethod
+    def backward(ctx, grad_output):
+        old_base, new_base = ctx.saved_tensors
+
+        packed_int = grad_output.to(torch.int64)
+        sign_bit = packed_int & 1
+        exponent = (packed_int >> 1).to(torch.float64)
+
+        exponent_new = exponent * torch.log(new_base) / torch.log(old_base)
+        new_packed_int = (exponent_new.round().to(torch.int64) << 1) | sign_bit
+        new_tensor = new_packed_int.to(torch.float64)
+
+        return new_tensor, None, None
+
+def align_lnstensor_bases(
+        *tensors: LNSTensor,
+        base: torch.Tensor | None = None
+    ) -> Tuple[LNSTensor, ...]:
+    """
+    Aligns the bases of a sequence of LNSTensors to a common base.
+
+    Parameters
+    ----------
+    tensors : LNSTensor
+        Variable number of LNSTensor objects to be aligned.
+    base : torch.Tensor, optional
+        The target base to which all tensors should be aligned.
+        If None, the default base from `xl.xlnsB` will be used.
+
+    Returns
+    -------
+    Tuple[LNSTensor, ...]
+        A tuple containing the LNSTensors with their bases aligned
+        to the specified base or default base. Tensors that already
+        match the base will be returned unchanged.
+
+    Notes
+    -----
+    This function ensures compatibility for operations requiring a
+    common logarithmic base. This operation is tracked by PyTorch's
+    autograd system to allow for correct gradient computation on the
+    original tensors in their original bases.
+    """
+    if base is None:
+        new_base = torch.tensor(xl.xlnsB, dtype=torch.float64)
+    else:
+        new_base = base.detach()
+
+    aligned_tensors = []
+    for tensor in tensors:
+
+        if torch.eq(tensor.base, new_base):
+            aligned_tensors.append(tensor)
+        else:
+            aligned_tensor = LNSChangeBaseFunction.apply(tensor, tensor.base, new_base)
+            aligned_tensors.append(lnstensor(aligned_tensor, from_lns=True, b=new_base))
+
+    return tuple(aligned_tensors)
+
+def format_lnstensor_operands(*operands: Any) -> Tuple[LNSTensor, ...]:
+    """
+    Converts a variable number of operands to LNSTensor objects, aligning
+    all operands to the base of the first operand that is an LNSTensor.
+
+    Parameters
+    ----------
+    operands : Any
+        Variable number of operands, which can be LNSTensor objects or
+        other array-like objects that can be converted to LNSTensor.
+
+    Returns
+    -------
+    Tuple[LNSTensor, ...]
+        A tuple of LNSTensor objects with their bases aligned to the base
+        of the first LNSTensor operand. If no LNSTensor is found, all
+        operands are converted to LNSTensors with the default base.
+    """
+    base = None
+
+    for operand in operands:
+        if isinstance(operand, LNSTensor):
+            base = operand.base
+            break
+    else:
+        base = torch.tensor(xl.xlnsB, dtype=torch.float64)
+
+    converted_operands = []
+    for operand in operands:
+        if isinstance(operand, LNSTensor):
+            converted_operands.append(operand)
+        else:
+            converted_operands.append(lnstensor(operand, detach=False, b=base))
+
+    return align_lnstensor_bases(*converted_operands, base=base)
 
 def lnstensor(
         data: Any,
